Pulled out the game modes from frontend
