+++ src/frontend.cpp
@@ -82,408 +82,9 @@
 #include "uiscreens/titlemenu.h"
 #include "uiscreens/common.h"
 #include "uiscreens/options/graphicsoptions.h"
-
-// ////////////////////////////////////////////////////////////////////////////
-// Global variables
-
-char			aLevelName[MAX_LEVEL_NAME_SIZE + 1];	//256];			// vital! the wrf file to use.
-
-bool			bLimiterLoaded = false;
-
-#define TUTORIAL_LEVEL "TUTORIAL3"
-
-// ////////////////////////////////////////////////////////////////////////////
-// Tutorial Menu
-void startTutorialMenu()
-{
-	addBackdrop();
-	addTopForm(false);
-	addBottomForm();
+#include "uiscreens/gamemodes/gmcommon.h"
 
 
-	addTextButton(FRONTEND_TUTORIAL, FRONTEND_POS3X, FRONTEND_POS3Y, _("Tutorial"), WBUT_TXTCENTRE);
-	addTextButton(FRONTEND_FASTPLAY, FRONTEND_POS4X, FRONTEND_POS4Y, _("Fast Play"), WBUT_TXTCENTRE);
-	addSideText(FRONTEND_SIDETEXT , FRONTEND_SIDEX, FRONTEND_SIDEY, _("TUTORIALS"));
-	// TRANSLATORS: "Return", in this context, means "return to previous screen/menu"
-	addMultiBut(psWScreen, FRONTEND_BOTFORM, FRONTEND_QUIT, 10, 10, 30, 29, P_("menu", "Return"), IMAGE_RETURN, IMAGE_RETURN_HI, IMAGE_RETURN_HI);
-}
-
-bool runTutorialMenu()
-{
-	WidgetTriggers const &triggers = widgRunScreen(psWScreen);
-	unsigned id = triggers.empty() ? 0 : triggers.front().widget->id; // Just use first click here, since the next click could be on another menu.
-
-	switch (id)
-	{
-	case FRONTEND_TUTORIAL:
-		SPinit(LEVEL_TYPE::CAMPAIGN);
-		sstrcpy(aLevelName, TUTORIAL_LEVEL);
-		changeTitleMode(STARTGAME);
-		break;
-
-	case FRONTEND_FASTPLAY:
-		SPinit(LEVEL_TYPE::CAMPAIGN);
-		sstrcpy(aLevelName, "FASTPLAY");
-		changeTitleMode(STARTGAME);
-		break;
-
-	case FRONTEND_QUIT:
-		changeTitleMode(TITLE);
-		break;
-
-	default:
-		break;
-	}
-
-	// If close button pressed then return from this menu.
-	if (CancelPressed())
-	{
-		changeTitleMode(TITLE);
-	}
-
-	widgDisplayScreen(psWScreen);						// show the widgets currently running
-
-	return true;
-}
-
-
-// ////////////////////////////////////////////////////////////////////////////
-// Single Player Menu
-void startSinglePlayerMenu()
-{
-	challengeActive = false;
-	addBackdrop();
-	addTopForm(false);
-	addBottomForm();
-
-	addTextButton(FRONTEND_NEWGAME,  FRONTEND_POS2X, FRONTEND_POS2Y, _("New Campaign") , WBUT_TXTCENTRE);
-	addTextButton(FRONTEND_SKIRMISH, FRONTEND_POS3X, FRONTEND_POS3Y, _("Start Skirmish Game"), WBUT_TXTCENTRE);
-	addTextButton(FRONTEND_CHALLENGES, FRONTEND_POS4X, FRONTEND_POS4Y, _("Challenges"), WBUT_TXTCENTRE);
-	addTextButton(FRONTEND_LOADGAME_MISSION, FRONTEND_POS5X, FRONTEND_POS5Y, _("Load Campaign Game"), WBUT_TXTCENTRE);
-	addTextButton(FRONTEND_LOADGAME_SKIRMISH, FRONTEND_POS6X, FRONTEND_POS6Y, _("Load Skirmish Game"), WBUT_TXTCENTRE);
-	addTextButton(FRONTEND_REPLAY, FRONTEND_POS7X,FRONTEND_POS7Y, _("View Skirmish Replay"), WBUT_TXTCENTRE);
-
-	addSideText(FRONTEND_SIDETEXT, FRONTEND_SIDEX, FRONTEND_SIDEY, _("SINGLE PLAYER"));
-	addMultiBut(psWScreen, FRONTEND_BOTFORM, FRONTEND_QUIT, 10, 10, 30, 29, P_("menu", "Return"), IMAGE_RETURN, IMAGE_RETURN_HI, IMAGE_RETURN_HI);
-	// show this only when the video sequences are not installed
-	if (!PHYSFS_exists("sequences/devastation.ogg"))
-	{
-		addSmallTextButton(FRONTEND_HYPERLINK, FRONTEND_POS9X, FRONTEND_POS9Y, _("Campaign videos are missing! Get them from http://wz2100.net"), 0);
-	}
-}
-
-void startCampaignSelector()
-{
-	addBackdrop();
-	addTopForm(false);
-	addBottomForm();
-
-	std::vector<CAMPAIGN_FILE> list = readCampaignFiles();
-	ASSERT(list.size() <= static_cast<size_t>(std::numeric_limits<UDWORD>::max()), "list.size() (%zu) exceeds UDWORD max", list.size());
-	for (UDWORD i = 0; i < static_cast<UDWORD>(list.size()); i++)
-	{
-		addTextButton(FRONTEND_CAMPAIGN_1 + i, FRONTEND_POS1X, FRONTEND_POS2Y + FRONTEND_BUTHEIGHT * i, gettext(list[i].name.toUtf8().c_str()), WBUT_TXTCENTRE);
-	}
-	addSideText(FRONTEND_SIDETEXT, FRONTEND_SIDEX, FRONTEND_SIDEY, _("CAMPAIGNS"));
-	addMultiBut(psWScreen, FRONTEND_BOTFORM, FRONTEND_QUIT, 10, 10, 30, 29, P_("menu", "Return"), IMAGE_RETURN, IMAGE_RETURN_HI, IMAGE_RETURN_HI);
-	// show this only when the video sequences are not installed
-	if (!PHYSFS_exists("sequences/devastation.ogg"))
-	{
-		addSmallTextButton(FRONTEND_HYPERLINK, FRONTEND_POS9X, FRONTEND_POS9Y, _("Campaign videos are missing! Get them from http://wz2100.net"), 0);
-	}
-}
-
-static void frontEndNewGame(int which)
-{
-	std::vector<CAMPAIGN_FILE> list = readCampaignFiles();
-	sstrcpy(aLevelName, list[which].level.toUtf8().c_str());
-	// show this only when the video sequences are installed
-	if (PHYSFS_exists("sequences/devastation.ogg"))
-	{
-		if (!list[which].video.isEmpty())
-		{
-			seq_ClearSeqList();
-			seq_AddSeqToList(list[which].video.toUtf8().c_str(), nullptr, list[which].captions.toUtf8().c_str(), false);
-			seq_StartNextFullScreenVideo();
-		}
-	}
-	if (!list[which].package.isEmpty())
-	{
-		WzString path;
-		path += PHYSFS_getWriteDir();
-		path += PHYSFS_getDirSeparator();
-		path += "campaigns";
-		path += PHYSFS_getDirSeparator();
-		path += list[which].package;
-		if (!PHYSFS_mount(path.toUtf8().c_str(), NULL, PHYSFS_APPEND))
-		{
-			debug(LOG_ERROR, "Failed to load campaign mod \"%s\": %s",
-			      path.toUtf8().c_str(), WZ_PHYSFS_getLastError());
-		}
-	}
-	if (!list[which].loading.isEmpty())
-	{
-		debug(LOG_WZ, "Adding campaign mod level \"%s\"", list[which].loading.toUtf8().c_str());
-		if (!loadLevFile(list[which].loading.toUtf8().c_str(), mod_campaign, false, nullptr))
-		{
-			debug(LOG_ERROR, "Failed to load %s", list[which].loading.toUtf8().c_str());
-			return;
-		}
-	}
-	debug(LOG_WZ, "Loading campaign mod -- %s", aLevelName);
-	changeTitleMode(STARTGAME);
-}
-
-static void loadOK()
-{
-	if (strlen(sRequestResult))
-	{
-		sstrcpy(saveGameName, sRequestResult);
-		changeTitleMode(LOADSAVEGAME);
-	}
-}
-
-void SPinit(LEVEL_TYPE gameType)
-{
-	uint8_t playercolor;
-
-	NetPlay.bComms = false;
-	bMultiPlayer = false;
-	bMultiMessages = false;
-	game.type = gameType;
-	NET_InitPlayers();
-	NetPlay.players[0].allocated = true;
-	NetPlay.players[0].autoGame = false;
-	NetPlay.players[0].difficulty = AIDifficulty::HUMAN;
-	game.maxPlayers = MAX_PLAYERS - 1;	// Currently, 0 - 10 for a total of MAX_PLAYERS
-	// make sure we have a valid color choice for our SP game. Valid values are 0, 4-7
-	playercolor = war_GetSPcolor();
-
-	if (playercolor >= 1 && playercolor <= 3)
-	{
-		playercolor = 0;	// default is green
-	}
-	setPlayerColour(0, playercolor);
-	game.hash.setZero();	// must reset this to zero
-	builtInMap = true;
-}
-
-bool runCampaignSelector()
-{
-	WidgetTriggers const &triggers = widgRunScreen(psWScreen);
-	unsigned id = triggers.empty() ? 0 : triggers.front().widget->id; // Just use first click here, since the next click could be on another menu.
-	if (id == FRONTEND_QUIT)
-	{
-		changeTitleMode(SINGLE); // go back
-	}
-	else if (id >= FRONTEND_CAMPAIGN_1 && id <= FRONTEND_CAMPAIGN_6) // chose a campaign
-	{
-		SPinit(LEVEL_TYPE::CAMPAIGN);
-		frontEndNewGame(id - FRONTEND_CAMPAIGN_1);
-	}
-	else if (id == FRONTEND_HYPERLINK)
-	{
-		runHyperlink();
-	}
-
-	widgDisplayScreen(psWScreen); // show the widgets currently running
-
-	if (CancelPressed())
-	{
-		changeTitleMode(SINGLE);
-	}
-
-	return true;
-}
-
-bool runSinglePlayerMenu()
-{
-	if (bLoadSaveUp)
-	{
-		if (runLoadSave(false)) // check for file name.
-		{
-			loadOK();
-		}
-	}
-	else if (challengesUp)
-	{
-		runChallenges();
-	}
-	else
-	{
-		WidgetTriggers const &triggers = widgRunScreen(psWScreen);
-		unsigned id = triggers.empty() ? 0 : triggers.front().widget->id; // Just use first click here, since the next click could be on another menu.
-
-		switch (id)
-		{
-		case FRONTEND_NEWGAME:
-			changeTitleMode(CAMPAIGNS);
-			break;
-
-		case FRONTEND_LOADGAME_MISSION:
-			SPinit(LEVEL_TYPE::CAMPAIGN);
-			addLoadSave(LOAD_FRONTEND_MISSION, _("Load Campaign Saved Game"));	// change mode when loadsave returns
-			break;
-
-		case FRONTEND_LOADGAME_SKIRMISH:
-			SPinit(LEVEL_TYPE::SKIRMISH);
-			bMultiPlayer = true;
-			addLoadSave(LOAD_FRONTEND_SKIRMISH, _("Load Skirmish Saved Game"));	// change mode when loadsave returns
-			break;
-
-		case FRONTEND_REPLAY:
-			SPinit(LEVEL_TYPE::SKIRMISH);
-			bMultiPlayer = true;
-			addLoadSave(LOADREPLAY_FRONTEND_SKIRMISH, _("Load Skirmish Replay"));  // change mode when loadsave returns
-			break;
-
-		case FRONTEND_SKIRMISH:
-			SPinit(LEVEL_TYPE::SKIRMISH);
-			sstrcpy(game.map, DEFAULTSKIRMISHMAP);
-			game.hash = levGetMapNameHash(game.map);
-			game.maxPlayers = DEFAULTSKIRMISHMAPMAXPLAYERS;
-
-			ingame.side = InGameSide::HOST_OR_SINGLEPLAYER;
-			changeTitleUI(std::make_shared<WzMultiplayerOptionsTitleUI>(wzTitleUICurrent));
-			break;
-
-		case FRONTEND_QUIT:
-			changeTitleMode(TITLE);
-			break;
-
-		case FRONTEND_CHALLENGES:
-			SPinit(LEVEL_TYPE::SKIRMISH);
-			addChallenges();
-			break;
-
-		case FRONTEND_HYPERLINK:
-			runHyperlink();
-			break;
-
-		default:
-			break;
-		}
-
-		if (CancelPressed())
-		{
-			changeTitleMode(TITLE);
-		}
-	}
-
-	if (!bLoadSaveUp && !challengesUp)						// if save/load screen is up
-	{
-		widgDisplayScreen(psWScreen);						// show the widgets currently running
-	}
-	if (bLoadSaveUp)								// if save/load screen is up
-	{
-		displayLoadSave();
-	}
-	else if (challengesUp)
-	{
-		displayChallenges();
-	}
-
-	return true;
-}
-
-
-// ////////////////////////////////////////////////////////////////////////////
-// Multi Player Menu
-void startMultiPlayerMenu()
-{
-	addBackdrop();
-	addTopForm(false);
-	addBottomForm();
-
-	addSideText(FRONTEND_SIDETEXT ,	FRONTEND_SIDEX, FRONTEND_SIDEY, _("MULTI PLAYER"));
-
-	addTextButton(FRONTEND_HOST,     FRONTEND_POS2X, FRONTEND_POS2Y, _("Host Game"), WBUT_TXTCENTRE);
-	addTextButton(FRONTEND_JOIN,     FRONTEND_POS3X, FRONTEND_POS3Y, _("Join Game"), WBUT_TXTCENTRE);
-	addTextButton(FRONTEND_REPLAY,   FRONTEND_POS7X, FRONTEND_POS7Y, _("View Replay"), WBUT_TXTCENTRE);
-
-	addMultiBut(psWScreen, FRONTEND_BOTFORM, FRONTEND_QUIT, 10, 10, 30, 29, P_("menu", "Return"), IMAGE_RETURN, IMAGE_RETURN_HI, IMAGE_RETURN_HI);
-
-	// This isn't really a hyperlink for now... perhaps link to the wiki ?
-	addSmallTextButton(FRONTEND_HYPERLINK, FRONTEND_POS9X, FRONTEND_POS9Y, _("TCP port 2100 must be opened in your firewall or router to host games!"), 0);
-}
-
-bool runMultiPlayerMenu()
-{
-	if (bLoadSaveUp)
-	{
-		if (runLoadSave(false)) // check for file name.
-		{
-			loadOK();
-		}
-	}
-	else
-	{
-		WidgetTriggers const &triggers = widgRunScreen(psWScreen);
-		unsigned id = triggers.empty() ? 0 : triggers.front().widget->id; // Just use first click here, since the next click could be on another menu.
-
-		switch (id)
-		{
-		case FRONTEND_HOST:
-			// don't pretend we are running a network game. Really do it!
-			NetPlay.bComms = true; // use network = true
-			NetPlay.isUPNP_CONFIGURED = false;
-			NetPlay.isUPNP_ERROR = false;
-			ingame.side = InGameSide::HOST_OR_SINGLEPLAYER;
-			bMultiPlayer = true;
-			bMultiMessages = true;
-			NETinit(true);
-			NETdiscoverUPnPDevices();
-			game.type = LEVEL_TYPE::SKIRMISH;		// needed?
-			changeTitleUI(std::make_shared<WzMultiplayerOptionsTitleUI>(wzTitleUICurrent));
-			break;
-		case FRONTEND_JOIN:
-			NETinit(true);
-			ingame.side = InGameSide::MULTIPLAYER_CLIENT;
-			if (getLobbyError() != ERROR_INVALID)
-			{
-				setLobbyError(ERROR_NOERROR);
-			}
-			changeTitleUI(std::make_shared<WzProtocolTitleUI>());
-			break;
-		case FRONTEND_REPLAY:
-			SPinit(LEVEL_TYPE::SKIRMISH);
-			bMultiPlayer = true;
-			game.maxPlayers = DEFAULTSKIRMISHMAPMAXPLAYERS;
-			addLoadSave(LOADREPLAY_FRONTEND_MULTI, _("Load Multiplayer Replay"));  // change mode when loadsave returns
-			break;
-
-		case FRONTEND_QUIT:
-			changeTitleMode(TITLE);
-			break;
-		default:
-			break;
-		}
-
-		if (CancelPressed())
-		{
-			changeTitleMode(TITLE);
-		}
-	}
-
-	if (!bLoadSaveUp)
-	{
-		widgDisplayScreen(psWScreen);		// show the widgets currently running
-	}
-	else if (bLoadSaveUp)					// if save/load screen is up
-	{
-		displayLoadSave();
-	}
-
-	return true;
-}
-
-
-// ////////////////////////////////////////////////////////////////////////////
-// Change Mode
-void changeTitleMode(tMode mode)
-{
-	changeTitleUI(std::make_shared<WzOldTitleUI>(mode));
-}
 
 // ////////////////////////////////////////////////////////////////////////////
 // Handling Screen Size Changes
