+++ src/multilimit.cpp
@@ -55,6 +55,7 @@
 #include "titleui/titleui.h"
 
 #include "uiscreens/common.h"
+#include "uiscreens/gamemodes/gmcommon.h"
 
 // ////////////////////////////////////////////////////////////////////////////
 // defines
