+++ src/uiscreens/gamemodes/campaignselector.h
@@ -1,0 +1,7 @@
+#pragma once
+
+#include "../common.h"
+#include "gmcommon.h"
+
+bool runCampaignSelector();
+void startCampaignSelector();
