+++ src/uiscreens/gamemodes/multiplayermenu.h
@@ -1,0 +1,7 @@
+#pragma once
+
+#include "../common.h"
+#include "gmcommon.h"
+
+bool runMultiPlayerMenu();
+void startMultiPlayerMenu();
