+++ src/uiscreens/gamemodes/singleplayermenu.h
@@ -1,0 +1,7 @@
+#pragma once
+
+#include "../common.h"
+#include "gmcommon.h"
+
+bool runSinglePlayerMenu();
+void startSinglePlayerMenu();
