+++ src/uiscreens/gamemodes/tutorialmenu.cpp
@@ -1,0 +1,55 @@
+#include "tutorialmenu.h"
+
+// ////////////////////////////////////////////////////////////////////////////
+// Tutorial Menu
+void startTutorialMenu()
+{
+	addBackdrop();
+	addTopForm(false);
+	addBottomForm();
+
+
+	addTextButton(FRONTEND_TUTORIAL, FRONTEND_POS3X, FRONTEND_POS3Y, _("Tutorial"), WBUT_TXTCENTRE);
+	addTextButton(FRONTEND_FASTPLAY, FRONTEND_POS4X, FRONTEND_POS4Y, _("Fast Play"), WBUT_TXTCENTRE);
+	addSideText(FRONTEND_SIDETEXT, FRONTEND_SIDEX, FRONTEND_SIDEY, _("TUTORIALS"));
+	// TRANSLATORS: "Return", in this context, means "return to previous screen/menu"
+	addMultiBut(psWScreen, FRONTEND_BOTFORM, FRONTEND_QUIT, 10, 10, 30, 29, P_("menu", "Return"), IMAGE_RETURN, IMAGE_RETURN_HI, IMAGE_RETURN_HI);
+}
+
+bool runTutorialMenu()
+{
+	WidgetTriggers const& triggers = widgRunScreen(psWScreen);
+	unsigned id = triggers.empty() ? 0 : triggers.front().widget->id; // Just use first click here, since the next click could be on another menu.
+
+	switch (id)
+	{
+	case FRONTEND_TUTORIAL:
+		SPinit(LEVEL_TYPE::CAMPAIGN);
+		sstrcpy(aLevelName, TUTORIAL_LEVEL);
+		changeTitleMode(STARTGAME);
+		break;
+
+	case FRONTEND_FASTPLAY:
+		SPinit(LEVEL_TYPE::CAMPAIGN);
+		sstrcpy(aLevelName, "FASTPLAY");
+		changeTitleMode(STARTGAME);
+		break;
+
+	case FRONTEND_QUIT:
+		changeTitleMode(TITLE);
+		break;
+
+	default:
+		break;
+	}
+
+	// If close button pressed then return from this menu.
+	if (CancelPressed())
+	{
+		changeTitleMode(TITLE);
+	}
+
+	widgDisplayScreen(psWScreen);						// show the widgets currently running
+
+	return true;
+}
