+++ src/uiscreens/gamemodes/tutorialmenu.h
@@ -1,0 +1,10 @@
+#pragma once
+
+#include "../common.h"
+#include "gmcommon.h"
+
+#define TUTORIAL_LEVEL "TUTORIAL3"
+
+
+bool runTutorialMenu();
+void startTutorialMenu();
